+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Sample Junit5ModelGenerator
+
 ## [1.4.0] - 2024-02-27
 
 ### Added